firewall enrollment workflow (wraps enroll-netdev + management/mate metadata)
